add order and add review page created

+++ src/Pages/ManageOrders/OrderTable/OrderTable.js
@@ -13,8 +13,10 @@ const OrderTable = ({ myOrder }) => {
             })
     }, []);
     //cancel an order
+    
     const handleCancelOrder = (id) => {
-        fetch(`https://serene-beyond-56458.herokuapp.com/deleteorder/${id}`, {
+        if(window.confirm("Are you sure want to cancel this order?")){
+            fetch(`https://serene-beyond-56458.herokuapp.com/deleteorder/${id}`, {
             method: 'delete'
         })
             .then(res => res.json())
@@ -24,6 +26,10 @@ const OrderTable = ({ myOrder }) => {
                     window.location.reload();
                 }
             })
+        }else{
+            alert('Order not cancelled')
+        }
+        
     }
 
     //update order status
